feat(freecache): race问题
