+++ pkg/cache/xfreecache/v2/encode.go
@@ -18,6 +18,23 @@ func marshal[T any](cacheData T) (data []byte, err error) {
 	return
 }
 
+// 反序列化，如果是pb格式，则使用proto序列化
+func unmarshal[T any](body []byte) (value T, err error) {
+	if msg, ok := any(value).(proto.Message); ok { // Constrained to proto.Message
+		// Peek the type inside T (as T= *SomeProtoMsgType)
+		msgType := reflect.TypeOf(msg).Elem()
+
+		// Make a new one, and throw it back into T
+		msg = reflect.New(msgType).Interface().(proto.Message)
+
+		err = proto.Unmarshal(body, msg)
+		value = msg.(T)
+	} else {
+		err = json.Unmarshal(body, &value)
+	}
+	return
+}
+
 var pools sync.Map
 
 func getPool[T any]() *sync.Pool {
@@ -41,7 +58,7 @@ func getPool[T any]() *sync.Pool {
 	return nil
 }
 
-// 反序列化，如果是pb格式，则使用proto序列化 使用sync.Pool
+// 反序列化，如果是pb格式，则使用proto序列化 使用sync.Pool-存在并发问题
 func unmarshalWithPool[T any](body []byte, pool *sync.Pool) (value T, err error) {
 	if _, ok := any(value).(proto.Message); ok { // Constrained to proto.Message
 		msg := pool.Get().(proto.Message)
