+++ pkg/cache/xfreecache/v2/encode_bench_test.go
@@ -2,7 +2,6 @@ package xfreecache
 
 import (
 	"encoding/json"
-	"reflect"
 	"testing"
 
 	helloworldv1 "github.com/douyu/jupiter/proto/helloworld/v1"
@@ -20,7 +19,7 @@ var helloReply = &helloworldv1.SayHiResponse{
 }
 
 /*
-   encoding/json
+encoding/json
 */
 func BenchmarkDecodeStdStructMedium(b *testing.B) {
 	res, _ := json.Marshal(helloReply)
@@ -92,21 +91,4 @@ func BenchmarkDecodeProtoWithReflectAndPool(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		_, _ = unmarshalWithPool[*helloworldv1.SayHiResponse](res, pool)
 	}
-}
-
-// 反序列化，如果是pb格式，则使用proto序列化
-func unmarshal[T any](body []byte) (value T, err error) {
-	if msg, ok := any(value).(proto.Message); ok { // Constrained to proto.Message
-		// Peek the type inside T (as T= *SomeProtoMsgType)
-		msgType := reflect.TypeOf(msg).Elem()
-
-		// Make a new one, and throw it back into T
-		msg = reflect.New(msgType).Interface().(proto.Message)
-
-		err = proto.Unmarshal(body, msg)
-		value = msg.(T)
-	} else {
-		err = json.Unmarshal(body, &value)
-	}
-	return
 }
